test: 更新 Tab 单元测试
