+++ test/UnitTest/Components/TabTest.cs
@@ -9,8 +9,20 @@ using UnitTest.Misc;
 
 namespace UnitTest.Components;
 
-public class TabTest : TabTestBase
+public class TabTest : BootstrapBlazorTestBase
 {
+    protected override void ConfigureServices(IServiceCollection services)
+    {
+        services.AddBootstrapBlazor(op => op.ToastDelay = 2000);
+        services.ConfigureTabItemMenuBindOptions(options =>
+        {
+            options.Binders = new()
+            {
+                { "/Binder", new() { Text = "Index_Binder_Test" } }
+            };
+        });
+    }
+
     [Fact]
     public void TabItem_Ok()
     {
